Arena allocator code makeup
Move code from squtils.h to separate source file - it is only used
by AST compiler, so no need to pull it into every other source.
Especially heavyweight platform-specific headers with lots of macros.

Make the use of system memory allocation optional and disabled by
default - use overridable memory management functions (sq_vm_malloc/
sq_vm_realloc/sq_vm_free).
Since the VM is embeddable, we want host application to control
memory operations. This also provides portability.

Note: if OS API will be used, probably allocation parameters should
be adjusted (see dwAllocationGranularity on Windows).
